+++ lib/logstash/outputs/google_cloud_storage.rb
@@ -20,8 +20,10 @@
 # -----
 require "logstash/outputs/base"
 require "logstash/outputs/gcs/path_factory"
+require "logstash/outputs/gcs/worker_pool"
 require "logstash/namespace"
 require "logstash/json"
+require "stud/interval"
 require "zlib"
 
 # Summary: plugin to upload log events to Google Cloud Storage (GCS), rolling
@@ -71,8 +73,6 @@ require "zlib"
 # * There's no recover method, so if logstash/plugin crashes, files may not
 # be uploaded to GCS.
 # * Allow user to configure file name.
-# * Allow parallel uploads for heavier loads (+ connection configuration if
-# exposed by Ruby API client)
 class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
   config_name "google_cloud_storage"
 
@@ -137,28 +137,22 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
   # When this feature is enabled, the uploader_interval_secs option has no effect.
   config :upload_synchronous, :validate => :boolean, :default => false
 
-  public
+  config :max_concurrent_uploads, :validate  => :number, :default => 5
+
   def register
     require "fileutils"
     require "thread"
-
     @logger.debug("GCS: register plugin")
-
     @last_flush_cycle = Time.now
 
-    unless upload_synchronous
-      initialize_upload_queue()
-    end
-
-    initialize_temp_directory()
+    @workers = LogStash::Outputs::Gcs::WorkerPool.new(@max_concurrent_uploads, @upload_synchronous)
+    initialize_temp_directory
     initialize_path_factory
     open_current_file
 
-    initialize_google_client()
+    initialize_google_client
 
-    unless upload_synchronous
-      @uploader = start_uploader
-    end
+    start_uploader
 
     if @gzip
       @content_type = 'application/gzip'
@@ -169,7 +163,6 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
 
   # Method called for each log event. It writes the event to the current output
   # file, flushing depending on flush interval configuration.
-  public
   def receive(event)
     @logger.debug("GCS: receive method called", :event => event)
 
@@ -180,22 +173,7 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
     end
 
     # Time to roll file based on the date pattern? Or is it over the size limit?
-    if (@path_factory.should_rotate? || (@max_file_size_kbytes > 0 && @temp_file.size >= @max_file_size_kbytes * 1024))
-      @logger.debug("GCS: log file will be closed and uploaded",
-                    :filename => File.basename(@temp_file.to_path),
-                    :size => @temp_file.size.to_s,
-                    :max_size => @max_file_size_kbytes.to_s)
-      # Close does not guarantee that data is physically written to disk.
-      @temp_file.fsync()
-      @temp_file.close()
-
-      if upload_synchronous
-        upload_object(@temp_file.to_path)
-        File.delete(@temp_file.to_path)
-      end
-
-      initialize_next_log()
-    end
+    initialize_next_log if ready_to_rotate?
 
     @temp_file.write(message)
     @temp_file.write("\n")
@@ -206,23 +184,25 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
                   :filename => File.basename(@temp_file.to_path))
   end
 
-  public
   def close
-    @logger.debug("GCS: close method called")
+    @logger.debug('Stopping the plugin, uploading the remaining files.')
 
-    @temp_file.fsync()
-    filename = @temp_file.to_path
-    size = @temp_file.size
-    @temp_file.close()
+    Stud.stop!(@registration_thread) unless @registration_thread.nil?
 
-    if upload_synchronous && size > 0
-      @logger.debug("GCS: uploading last file of #{size.to_s}b")
-      upload_object(filename)
-      File.delete(filename)
-    end
+    close_and_upload_current
+    @workers.stop!
   end
 
   private
+
+
+  def ready_to_rotate?
+    path_changed = @path_factory.should_rotate?
+    too_big = @max_file_size_kbytes > 0 && @temp_file.size >= @max_file_size_kbytes * 1024
+
+    path_changed || too_big
+  end
+
   ##
   # Flushes temporary log file every flush_interval_secs seconds or so.
   # This is triggered by events, but if there are no events there's no point
@@ -276,69 +256,23 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
 
   def start_uploader
     Thread.new do
-      @logger.debug("GCS: starting uploader")
-      while true
-        upload_from_queue()
+      @registration_thread = Thread.current
+      Stud.interval(@uploader_interval_secs) do
+        initialize_next_log if ready_to_rotate?
       end
     end
-  end
-  ##
-  # Uploads log files.
-  #
-  # Uploader is done in a separate thread, not holding the receive method above.
-  def upload_from_queue
-    filename = @upload_queue.pop
-
-    # Reenqueue if it is still the current file.
-    if filename == @temp_file.to_path
-      if !@path_factory.should_rotate?
-        @logger.debug("GCS: reenqueue as log file is being currently appended to.",
-                      :filename => filename)
-        @upload_queue << filename
-        # If we got here, it means that older files were uploaded, so let's
-        # wait another minute before checking on this file again.
-        sleep @uploader_interval_secs
-        return
-      else
-        @logger.debug("GCS: flush and close file to be uploaded.",
-                      :filename => filename)
-        @temp_file.fsync()
-        @temp_file.close()
-        initialize_next_log()
-      end
-    end
-
-    if File.stat(filename).size > 0
-      upload_object(filename)
-    else
-      @logger.debug("GCS: file size is zero, skip upload.",
-                     :filename => filename,
-                     :filesize => File.stat(filename).size)
-    end
-    @logger.debug("GCS: delete local temporary file ",
-                  :filename => filename)
-    File.delete(filename)
   end
 
   ##
   # Opens current log file and updates @temp_file with an instance of IOWriter.
   # This method also adds file to the upload queue.
-  def open_current_file()
+  def open_current_file
     path = @path_factory.current_path
 
-    stat = File.stat(path) rescue nil
-    if stat and stat.ftype == "fifo" and RUBY_PLATFORM == "java"
-      fd = java.io.FileWriter.new(java.io.File.new(path))
-    else
-      fd = File.new(path, "a")
-    end
-    if @gzip
-      fd = Zlib::GzipWriter.new(fd)
-    end
+    fd = File.new(path, "a")
+    fd = Zlib::GzipWriter.new(fd) if @gzip
+
     @temp_file = GCSIOWriter.new(fd)
-    unless upload_synchronous
-      @upload_queue << @temp_file.to_path
-    end
   end
 
   ##
@@ -347,8 +281,9 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
   # the base log file name is the same (e.g. log file was not rolled given the
   # date pattern).
   def initialize_next_log
+    close_and_upload_current
     @path_factory.rotate_path!
-    open_current_file()
+    open_current_file
   end
 
   ##
@@ -367,15 +302,6 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
                                                          'https://www.googleapis.com/auth/devstorage.read_write',
                                                          key)
     @client.authorization = service_account.authorize
-  end
-
-  # Initialize the queue that harbors files to be uploaded
-  def initialize_upload_queue
-    @upload_queue = new_upload_queue()
-  end
-
-  def new_upload_queue
-    Queue.new
   end
 
   ##
@@ -403,6 +329,32 @@ class LogStash::Outputs::GoogleCloudStorage < LogStash::Outputs::Base
       sleep 1
       retry
     end
+  end
+
+  def close_and_upload_current
+    return if @temp_file.nil?
+
+    filename = @temp_file.to_path
+    @temp_file.fsync
+    @temp_file.close
+    @logger.info("Uploading file: #{filename}")
+
+    @workers.post do
+      upload_and_delete(filename)
+    end
+  end
+
+  def upload_and_delete(filename)
+    file_size = File.stat(filename).size
+
+    if file_size > 0
+      upload_object(filename)
+    else
+      @logger.debug('File size is zero, skip upload.', :filename => filename)
+    end
+
+    @logger.debug('Delete local temporary file', :filename => filename)
+    File.delete(filename)
   end
 end